vision
Implemented eye movement detection using OpenCV in Python. The program captures video from the webcam, detects faces and eyes, and tracks eye movements. It displays a message if eyes are closed or not looking at the screen and initiates a countdown timer. Upon reaching zero, it captures and saves a picture with a timestamp. Added functionality to continuously capture pictures every 2 seconds until eyes are detected back.